requirements.txt
PyYAML>=6.0
pdfplumber>=0.9.0
pytest>=7.0.0
